# world-game-app/packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Difficulty = typeof Difficulty[keyof typeof Difficulty];


export const Difficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type QuestionType = typeof QuestionType[keyof typeof QuestionType];


export const QuestionType = {
  flag: 'flag',
  capital: 'capital',
  continent: 'continent',
  currency: 'currency',
  dish: 'dish',
  language: 'language',
  tld: 'tld',
  government: 'government',
  independence: 'independence',
  symbol: 'symbol',
  density: 'density',
  religion: 'religion',
} as const;

export interface Question {
  id: string;
  difficulty: Difficulty;
  level: number;
  type: QuestionType;
  prompt: string;
  /**
     * Data URI of a country flag, present only for "flag" type questions
     * @nullable
     */
  flagImage: string | null;
  /**
     * @minItems 4
     * @maxItems 4
     */
  options: string[];
  /** Opaque signed token encoding the correct answer; required to verify or use a lifeline on this question */
  token: string;
}

export interface AnswerSubmission {
  token: string;
  /**
     * @minimum 0
     * @maximum 3
     */
  selectedIndex: number;
}

export interface AnswerResult {
  correct: boolean;
  correctIndex: number;
  correctAnswer: string;
}

export interface FiftyFiftyRequest {
  token: string;
}

export interface FiftyFiftyResult {
  /**
     * @minItems 2
     * @maxItems 2
     */
  removeIndices: number[];
}

export interface AskAudienceRequest {
  token: string;
}

export interface AskAudienceResult {
  /**
     * Simulated audience vote percentage for each option, in option order; sums to 100.
     * @minItems 4
     * @maxItems 4
     */
  percentages: number[];
}

export interface AskExpertRequest {
  token: string;
}

export interface AskExpertResult {
  /** The index the expert recommends, always the correct answer. */
  correctIndex: number;
}

export interface ScoreInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  playerName: string;
  /** @minimum 0 */
  winnings: number;
  difficulty: Difficulty;
  /** @minimum 0 */
  questionsAnswered: number;
  /** @minimum 0 */
  correctAnswers: number;
  won: boolean;
}

export interface Score {
  id: number;
  playerName: string;
  winnings: number;
  difficulty: Difficulty;
  questionsAnswered: number;
  correctAnswers: number;
  won: boolean;
  createdAt: string;
}

export type GetQuestionParams = {
difficulty: Difficulty;
/**
 * @minimum 1
 * @maximum 15
 */
level: number;
};

export type ListTopScoresParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};