# world-game-app/packages/api-spec/src/generated/zod/types/answerResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerResult {
  correct: boolean;
  correctIndex: number;
  correctAnswer: string;
}