# world-game-app/packages/api-spec/src/generated/zod/types/askAudienceResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AskAudienceResult {
  /**
     * Simulated audience vote percentage for each option, in option order; sums to 100.
     * @minItems 4
     * @maxItems 4
     */
  percentages: number[];
}