# world-game-app/packages/api-spec/src/generated/zod/types/difficulty.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Difficulty = typeof Difficulty[keyof typeof Difficulty];


export const Difficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;