# world-game-app/packages/api-spec/src/generated/zod/types/fiftyFiftyResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FiftyFiftyResult {
  /**
     * @minItems 2
     * @maxItems 2
     */
  removeIndices: number[];
}