# world-game-app/packages/api-spec/src/generated/zod/types/questionType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type QuestionType = typeof QuestionType[keyof typeof QuestionType];


export const QuestionType = {
  flag: 'flag',
  capital: 'capital',
  continent: 'continent',
  currency: 'currency',
  dish: 'dish',
  language: 'language',
  tld: 'tld',
  government: 'government',
  independence: 'independence',
  symbol: 'symbol',
  density: 'density',
  religion: 'religion',
} as const;