# world-game-app/packages/api-spec/src/generated/zod/types/scoreInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Difficulty } from './difficulty.js';

export interface ScoreInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  playerName: string;
  /** @minimum 0 */
  winnings: number;
  difficulty: Difficulty;
  /** @minimum 0 */
  questionsAnswered: number;
  /** @minimum 0 */
  correctAnswers: number;
  won: boolean;
}