# world-game-app/packages/api-zod/src/generated/types/answerSubmission.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerSubmission {
  token: string;
  /**
     * @minimum 0
     * @maximum 3
     */
  selectedIndex: number;
}