# world-game-app/packages/api-zod/src/generated/types/askExpertResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AskExpertResult {
  /** The index the expert recommends, always the correct answer. */
  correctIndex: number;
}