# world-game-app/packages/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Difficulty } from './difficulty.js';
import type { QuestionType } from './questionType.js';

export interface Question {
  id: string;
  difficulty: Difficulty;
  level: number;
  type: QuestionType;
  prompt: string;
  /**
     * Data URI of a country flag, present only for "flag" type questions
     * @nullable
     */
  flagImage: string | null;
  /**
     * @minItems 4
     * @maxItems 4
     */
  options: string[];
  /** Opaque signed token encoding the correct answer; required to verify or use a lifeline on this question */
  token: string;
}